pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kube-ops"
dynamic = ["version"]
description = "Kubernetes OOP"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "myback", email = "54638513+myback@users.noreply.github.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "pyyaml~=6.0.1",
    "kubernetes~=30.1.0"
]

[project.urls]
"Homepage" = "https://github.com/myback/kube-ops"
"Bug Tracker" = "https://github.com/myback/kube-ops/issues"

[tool.hatch.version]
path = "src/kube_ops/__about__.py"

[tool.hatch.build.targets.sdist]
ignore-vcs = true
only-include = ["src/kube_ops"]

[tool.hatch.build.targets.wheel]
sources = ["src"]
